Replace all new occurrences of "long long" with "int64_t".

+++ tests/core/types/detail/get_last_if.cpp
@@ -41,24 +41,24 @@ void test_get_last_if() {
       "");
   static_assert(
       std::is_same<sqlpp::detail::get_last_if_t<sqlpp::is_integral, sqlpp::noop,
-                                                int, float, long long>,
-                   long long>::value,
+                                                int, float, int64_t>,
+                   int64_t>::value,
       "");
   static_assert(
       std::is_same<sqlpp::detail::get_last_if_t<sqlpp::is_integral, sqlpp::noop,
-                                                int, float, long long, short>,
+                                                int, float, int64_t, short>,
                    short>::value,
       "");
   static_assert(
       std::is_same<
           sqlpp::detail::get_last_if_t<sqlpp::is_integral, sqlpp::noop, int,
-                                       float, long long, short, size_t>,
+                                       float, int64_t, short, size_t>,
           short>::value,
       "");
   static_assert(
       std::is_same<
           sqlpp::detail::get_last_if_t<sqlpp::is_numeric, sqlpp::noop, int,
-                                       float, long long, short, size_t>,
+                                       float, int64_t, short, size_t>,
           size_t>::value,
       "");
 
@@ -80,13 +80,13 @@ void test_get_last_if() {
       "");
   static_assert(
       std::is_same<sqlpp::detail::get_last_if_t<sqlpp::is_integral, sqlpp::noop,
-                                                int, float, long long, double>,
-                   long long>::value,
+                                                int, float, int64_t, double>,
+                   int64_t>::value,
       "");
   static_assert(
       std::is_same<
           sqlpp::detail::get_last_if_t<sqlpp::is_integral, sqlpp::noop, int,
-                                       float, long long, short, double>,
+                                       float, int64_t, short, double>,
           short>::value,
       "");
 }
